restructing codebase into src/ packages
