+++ src/checkpointing/__init__.py
@@ -1,0 +1,13 @@
+"""
+Pico Checkpointing Package
+
+We subdivide the checkpointing into training and evaluation, and learning_dynamics. Training
+checkpoints store the model, optimizer, and learning rate scheduler. Evaluation checkpoints store
+the evaluation results. Learning dynamics checkpoints store activations and gradients used for
+learning dynamics analysis.
+"""
+
+# ruff: noqa: F401
+
+from .training import load_checkpoint, save_checkpoint
+from .evaluation import save_evaluation_results
