+++ src/model/__init__.py
@@ -1,0 +1,10 @@
+"""
+Model Package
+
+This Package contains the Pico model. If you have other models you'd like to implement, we
+recommend you add modules to this package.
+"""
+
+# ruff: noqa: F401
+
+from .pico import Pico
