--- utils/__init__.py
@@ -1,18 +1,0 @@
-"""
-Basic constants used by utility functions.
-"""
-
-# Root directory for storing training runs
-RUNS_DIR = "runs"
-
-# Subdirectory in RUNS_DIR for storing logs
-LOG_DIR = "logs"
-
-# Subdirectory in RUNS_DIR for storing evaluation results
-EVAL_RESULTS_DIR = "eval_results"
-
-# Subdirectory in RUNS_DIR for storing checkpoints
-CHECKPOINT_DIR = "checkpoints"
-
-# Subdirectory in RUNS_DIR for storing fabric checkpoint files
-FABRIC_CHECKPOINT_DIR = "fabric_state"
